refactor(decision-framework): migrate to code mode architecture with core, codemode, and mcp layers
- Restructure project into core, codemode, and mcp layers for better separation of concerns
- Extract core logic into DecisionManager, DecisionVisualizer, and type definitions
- Create DecisionFramework API wrapper in codemode layer for public consumption
- Rebuild MCP server to use the new API while maintaining existing functionality
- Add comprehensive test suites for each layer (core, codemode, mcp)
- Update build configuration to use separate tsconfig.build.json for type declarations
- Remove test-only accessor file in favor of integrated test utilities
