+++ .trae/documents/plan_20260212_172420.md
@@ -1,0 +1,39 @@
+# Migration Plan: Decision Framework to Code Mode Architecture
+
+I will migrate the `decision-framework` MCP server to the Code Mode architecture (Core / Code Mode / MCP) using the `mcp-code-architect` agent.
+
+## 1. Architecture Restructuring
+The monolithic `index.ts` will be split into three distinct layers:
+
+### Core Layer (`src/core/`)
+- **`types.ts`**: Will contain all interfaces (`Option`, `Outcome`, `Criterion`, `DecisionAnalysisData`, etc.).
+- **`manager.ts`**: Will contain the `DecisionManager` class responsible for:
+  - State management (`decisionHistory`, `registries`).
+  - Validation logic (`validateDecisionAnalysisData`).
+  - Calculation logic (`calculateExpectedValues`, `calculateMultiCriteriaScores`).
+- **`visualizer.ts`**: Will contain the `visualizeDecisionAnalysis` logic (using `chalk`) to keep the Manager pure.
+
+### Code Mode Layer (`src/codemode/`)
+- **`index.ts`**: Will export the `DecisionFrameworkAPI` class.
+  - This class will wrap `DecisionManager`.
+  - It will expose a typed method `analyze(data: DecisionAnalysisData): Promise<DecisionResult>`.
+
+### MCP Layer (`src/mcp/`)
+- **`tools.ts`**: Will contain the `DECISION_FRAMEWORK_TOOL` definition.
+- **`server.ts`**: Will contain the MCP server setup.
+  - It will instantiate `DecisionFrameworkAPI`.
+  - It will map the `decisionFramework` tool call to the API.
+  - It will use `visualizer` to format the output for the user.
+
+## 2. Implementation Steps
+1.  **Initialize Directory Structure**: Create `src/core`, `src/codemode`, `src/mcp`.
+2.  **Extract Core**: Move types and logic to `src/core`.
+3.  **Implement Code Mode**: Create the public API wrapper.
+4.  **Implement MCP Adapter**: Reconstruct the server using the new API.
+5.  **Update Entry Point**: Update `src/index.ts` to export the API and conditionally run the server.
+6.  **Migrate Tests**: specific tests for each layer.
+
+## 3. Validation
+- **Build**: Ensure `bun build` works and types are generated correctly.
+- **Test**: Ensure all existing tests pass (`bun test`).
+- **Functionality**: Verify that the tool still behaves exactly as before (input validation, calculation, visualization).
