+++ src/decision-framework/src/mcp/tools.ts
@@ -1,0 +1,271 @@
+import { Tool } from "@modelcontextprotocol/sdk/types.js";
+
+export const DECISION_FRAMEWORK_TOOL: Tool = {
+  name: "decisionFramework",
+  description: `A detailed tool for structured decision analysis and rational choice.
+This tool helps models systematically evaluate options, criteria, and outcomes.
+It supports multiple decision frameworks, probability estimates, and value judgments.
+
+When to use this tool:
+- Complex decisions with multiple options
+- Decisions requiring systematic evaluation of trade-offs
+- Decisions under uncertainty with probability estimates
+- Multi-stakeholder decisions with different criteria
+- Risk analysis and management
+
+Key features:
+- Structured option evaluation
+- Multi-criteria decision analysis
+- Expected utility calculations
+- Information value analysis
+- Visual decision support`,
+
+  inputSchema: {
+    type: "object",
+    additionalProperties: false,
+    properties: {
+      decisionStatement: {
+        type: "string",
+        description: "Clear statement of the decision to be made"
+      },
+      options: {
+        type: "array",
+        description: "Available options or alternatives",
+        items: {
+          type: "object",
+          additionalProperties: false,
+          properties: {
+            id: {
+              type: "string",
+              description: "Unique identifier for the option"
+            },
+            name: {
+              type: "string",
+              description: "Name of the option"
+            },
+            description: {
+              type: "string",
+              description: "Description of the option"
+            }
+          },
+          required: ["name", "description"]
+        }
+      },
+      criteria: {
+        type: "array",
+        description: "Criteria for evaluating options",
+        items: {
+          type: "object",
+          additionalProperties: false,
+          properties: {
+            id: {
+              type: "string",
+              description: "Unique identifier for the criterion"
+            },
+            name: {
+              type: "string",
+              description: "Name of the criterion"
+            },
+            description: {
+              type: "string",
+              description: "Description of the criterion"
+            },
+            weight: {
+              type: "number",
+              minimum: 0,
+              maximum: 1,
+              description: "Weight of the criterion (0.0-1.0)"
+            },
+            evaluationMethod: {
+              type: "string",
+              enum: ["quantitative", "qualitative", "boolean"],
+              description: "Method for evaluating options against this criterion"
+            }
+          },
+          required: ["name", "description", "weight", "evaluationMethod"]
+        }
+      },
+      criteriaEvaluations: {
+        type: "array",
+        description: "Evaluations of options against criteria",
+        items: {
+          type: "object",
+          additionalProperties: false,
+          properties: {
+            criterionId: {
+              type: "string",
+              description: "ID of the criterion"
+            },
+            optionId: {
+              type: "string",
+              description: "ID of the option"
+            },
+            score: {
+              type: "number",
+              minimum: 0,
+              maximum: 1,
+              description: "Score of the option on this criterion (0.0-1.0)"
+            },
+            justification: {
+              type: "string",
+              description: "Justification for the score"
+            }
+          },
+          required: ["criterionId", "optionId", "score", "justification"]
+        }
+      },
+      possibleOutcomes: {
+        type: "array",
+        description: "Possible outcomes and their probabilities",
+        items: {
+          type: "object",
+          additionalProperties: false,
+          properties: {
+            id: {
+              type: "string",
+              description: "Unique identifier for the outcome"
+            },
+            description: {
+              type: "string",
+              description: "Description of the outcome"
+            },
+            probability: {
+              type: "number",
+              minimum: 0,
+              maximum: 1,
+              description: "Probability of the outcome (0.0-1.0)"
+            },
+            optionId: {
+              type: "string",
+              description: "ID of the option this outcome is associated with"
+            },
+            value: {
+              type: "number",
+              description: "Utility value of the outcome (can be positive or negative)"
+            },
+            confidenceInEstimate: {
+              type: "number",
+              minimum: 0,
+              maximum: 1,
+              description: "Confidence in the probability and value estimates (0.0-1.0)"
+            }
+          },
+          required: ["description", "probability", "optionId", "value", "confidenceInEstimate"]
+        }
+      },
+      informationGaps: {
+        type: "array",
+        description: "Gaps in information that affect the decision",
+        items: {
+          type: "object",
+          additionalProperties: false,
+          properties: {
+            description: {
+              type: "string",
+              description: "Description of the information gap"
+            },
+            impact: {
+              type: "number",
+              minimum: 0,
+              maximum: 1,
+              description: "Impact of the information gap on the decision (0.0-1.0)"
+            },
+            researchMethod: {
+              type: "string",
+              description: "Method for filling the information gap"
+            }
+          },
+          required: ["description", "impact", "researchMethod"]
+        }
+      },
+      stakeholders: {
+        type: "array",
+        description: "People or groups affected by the decision",
+        items: {
+          type: "string"
+        }
+      },
+      constraints: {
+        type: "array",
+        description: "Constraints on the decision",
+        items: {
+          type: "string"
+        }
+      },
+      timeHorizon: {
+        type: "string",
+        description: "Time frame for the decision"
+      },
+      riskTolerance: {
+        type: "string",
+        enum: ["risk-averse", "risk-neutral", "risk-seeking"],
+        description: "Risk tolerance for the decision"
+      },
+      expectedValues: {
+        type: "object",
+        description: "Expected values for each option",
+        additionalProperties: {
+          type: "number"
+        }
+      },
+      multiCriteriaScores: {
+        type: "object",
+        description: "Multi-criteria scores for each option",
+        additionalProperties: {
+          type: "number"
+        }
+      },
+      sensitivityInsights: {
+        type: "array",
+        description: "Insights from sensitivity analysis",
+        items: {
+          type: "string"
+        }
+      },
+      recommendation: {
+        type: "string",
+        description: "Recommendation based on the analysis"
+      },
+      decisionId: {
+        type: "string",
+        description: "Unique identifier for this decision"
+      },
+      analysisType: {
+        type: "string",
+        enum: ["expected-utility", "multi-criteria", "maximin", "minimax-regret", "satisficing"],
+        description: "Type of decision analysis"
+      },
+      stage: {
+        type: "string",
+        enum: ["problem-definition", "options", "criteria", "evaluation", "analysis", "recommendation"],
+        description: "Current stage of the decision process"
+      },
+      iteration: {
+        type: "number",
+        minimum: 0,
+        description: "Current iteration of the decision analysis"
+      },
+      nextStageNeeded: {
+        type: "boolean",
+        description: "Whether another stage is needed"
+      },
+      suggestedNextStage: {
+        type: "string",
+        description: "Suggested next stage of the decision process"
+      }
+    },
+    required: [
+      "decisionStatement",
+      "options",
+      "stakeholders",
+      "constraints",
+      "timeHorizon",
+      "riskTolerance",
+      "decisionId",
+      "analysisType",
+      "stage",
+      "iteration",
+      "nextStageNeeded"
+    ]
+  }
+};
