+++ src/decision-framework/src/tests/core.test.ts
@@ -1,43 +1,17 @@
-import { describe, expect, it, beforeEach } from "bun:test";
-import createServer from "./index.js";
-import { CallToolRequestSchema } from "@modelcontextprotocol/sdk/types.js";
-import { validateDecisionAnalysisDataForTest } from "./test-accessors.js";
+import { describe, expect, it } from "bun:test";
+import { DecisionManager } from "../core/manager.js";
+import { DecisionAnalysisData } from "../core/types.js";
 
 /**
- * Test suite for Decision Framework MCP Server.
- *
- * Business Context: Ensures the decision framework correctly validates complex
- * decision analysis structures for enterprise strategic planning and risk management.
- *
- * Decision Rationale: Tests focus on server initialization, comprehensive validation,
- * and decision analysis logic to ensure production-ready reliability for strategic decisions.
+ * Testing-only accessor for validateDecisionAnalysisData method.
  */
-describe("Decision Framework Server", () => {
-  it("server initializes successfully", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-
-  it("server exports correct configuration", () => {
-    const server = createServer();
-    expect(typeof server.connect).toBe("function");
-    expect(typeof server.close).toBe("function");
-  });
-
-  it("server has correct name and version", () => {
-    const server = createServer();
-    expect(server).toBeDefined();
-  });
-});
+function validateDecisionAnalysisDataForTest(input: unknown) {
+  const manager = new DecisionManager();
+  return manager.validateDecisionAnalysisData(input);
+}
 
 /**
  * Decision Analysis Data Validation Tests.
- *
- * Business Context: Enterprise applications require comprehensive validation
- * of decision analysis data to ensure strategic decisions are based on accurate information.
- *
- * Decision Rationale: Test all validation paths including required fields,
- * enum values, and complex nested structures.
  */
 describe("Decision Analysis Data Validation", () => {
   it("validates complete decision analysis structure", () => {
@@ -131,7 +105,7 @@ describe("Decision Analysis Data Validation", () => {
       };
 
       const result = validateDecisionAnalysisDataForTest(input);
-      expect(result.analysisType).toBe(analysisType);
+      expect(result.analysisType).toBe(analysisType as DecisionAnalysisData["analysisType"]);
     }
   });
 
@@ -154,7 +128,7 @@ describe("Decision Analysis Data Validation", () => {
       };
 
       const result = validateDecisionAnalysisDataForTest(input);
-      expect(result.stage).toBe(stage);
+      expect(result.stage).toBe(stage as DecisionAnalysisData["stage"]);
     }
   });
 
@@ -177,7 +151,7 @@ describe("Decision Analysis Data Validation", () => {
       };
 
       const result = validateDecisionAnalysisDataForTest(input);
-      expect(result.riskTolerance).toBe(riskTolerance);
+      expect(result.riskTolerance).toBe(riskTolerance as DecisionAnalysisData["riskTolerance"]);
     }
   });
 
@@ -231,8 +205,8 @@ describe("Decision Analysis Data Validation", () => {
 
     const result = validateDecisionAnalysisDataForTest(inputWithEvaluations);
     expect(result.criteriaEvaluations).toHaveLength(2);
-    expect(result.criteriaEvaluations[0].score).toBe(0.7);
-    expect(result.criteriaEvaluations[0].justification).toBe("Lower cost than alternative");
+    expect(result.criteriaEvaluations![0].score).toBe(0.7);
+    expect(result.criteriaEvaluations![0].justification).toBe("Lower cost than alternative");
   });
 
   it("validates possible outcomes", () => {
@@ -262,8 +236,8 @@ describe("Decision Analysis Data Validation", () => {
 
     const result = validateDecisionAnalysisDataForTest(inputWithOutcomes);
     expect(result.possibleOutcomes).toHaveLength(1);
-    expect(result.possibleOutcomes[0].probability).toBe(0.8);
-    expect(result.possibleOutcomes[0].value).toBe(100000);
+    expect(result.possibleOutcomes![0].probability).toBe(0.8);
+    expect(result.possibleOutcomes![0].value).toBe(100000);
   });
 
   it("validates information gaps", () => {
@@ -291,8 +265,8 @@ describe("Decision Analysis Data Validation", () => {
 
     const result = validateDecisionAnalysisDataForTest(inputWithGaps);
     expect(result.informationGaps).toHaveLength(1);
-    expect(result.informationGaps[0].impact).toBe(0.7);
-    expect(result.informationGaps[0].researchMethod).toBe("Survey analysis");
+    expect(result.informationGaps![0].impact).toBe(0.7);
+    expect(result.informationGaps![0].researchMethod).toBe("Survey analysis");
   });
 
   it("validates analysis results", () => {
@@ -333,12 +307,6 @@ describe("Decision Analysis Data Validation", () => {
 
 /**
  * Input Validation Error Tests.
- *
- * Business Context: Enterprise applications require robust error handling
- * to prevent data corruption and ensure system stability.
- *
- * Decision Rationale: Test all validation error paths to ensure clear error
- * messages and proper input sanitization.
  */
 describe("Input Validation Errors", () => {
   it("rejects missing decisionStatement", () => {
@@ -487,15 +455,9 @@ describe("Input Validation Errors", () => {
 
 /**
  * Decision Processing Tests.
- *
- * Business Context: Decision processing must handle complex calculations
- * and provide accurate recommendations for enterprise decision-making.
- *
- * Decision Rationale: Test decision processing logic including expected value
- * calculations and multi-criteria analysis.
  */
 describe("Decision Processing", () => {
-  it("processes decision with expected utility analysis", async () => {
+  it("processes decision with expected utility analysis", () => {
     const input = {
       decisionStatement: "Investment decision",
       options: [{ id: "opt1", name: "High Risk Investment", description: "High risk, high reward" }],
@@ -528,15 +490,15 @@ describe("Decision Processing", () => {
       nextStageNeeded: false
     };
 
-    const { DecisionFrameworkServer } = await import("./index.js");
-    const serverInstance = new DecisionFrameworkServer();
-    const result = serverInstance.processDecisionAnalysis(input);
-    expect(result.content).toBeDefined();
-    expect(result.content[0].type).toBe("text");
-    expect(result.isError).toBeUndefined();
+    const manager = new DecisionManager();
+    const validated = manager.validateDecisionAnalysisData(input);
+    const result = manager.updateDecisionAnalysis(validated);
+
+    expect(result.expectedValues).toBeDefined();
+    expect(result.expectedValues?.opt1).toBeCloseTo(160000);
   });
 
-  it("processes decision with multi-criteria analysis", async () => {
+  it("processes decision with multi-criteria analysis", () => {
     const input = {
       decisionStatement: "Technology selection",
       options: [
@@ -597,244 +559,11 @@ describe("Decision Processing", () => {
       nextStageNeeded: false
     };
 
-    const { DecisionFrameworkServer } = await import("./index.js");
-    const serverInstance = new DecisionFrameworkServer();
-    const result = serverInstance.processDecisionAnalysis(input);
-    expect(result.content).toBeDefined();
-    expect(result.content[0].type).toBe("text");
-    expect(result.isError).toBeUndefined();
-  });
+    const manager = new DecisionManager();
+    const validated = manager.validateDecisionAnalysisData(input);
+    const result = manager.updateDecisionAnalysis(validated);
 
-  it("handles decision with information gaps", async () => {
-    const input = {
-      decisionStatement: "Market expansion decision",
-      options: [
-        { id: "opt1", name: "Expand to Europe", description: "European market expansion" },
-        { id: "opt2", name: "Expand to Asia", description: "Asian market expansion" }
-      ],
-      informationGaps: [
-        {
-          description: "European market research",
-          impact: 0.8,
-          researchMethod: "Market analysis"
-        },
-        {
-          description: "Regulatory compliance in Asia",
-          impact: 0.6,
-          researchMethod: "Legal review"
-        }
-      ],
-      stakeholders: ["CEO", "Legal Team"],
-      constraints: ["Limited budget"],
-      timeHorizon: "2 years",
-      riskTolerance: "risk-averse",
-      decisionId: "expansion-001",
-      analysisType: "multi-criteria",
-      stage: "problem-definition",
-      iteration: 1,
-      nextStageNeeded: true,
-      suggestedNextStage: "criteria"
-    };
-
-    const { DecisionFrameworkServer } = await import("./index.js");
-    const serverInstance = new DecisionFrameworkServer();
-    const result = serverInstance.processDecisionAnalysis(input);
-    expect(result.content).toBeDefined();
-    expect(result.content[0].type).toBe("text");
-    expect(result.isError).toBeUndefined();
-  });
-});
-
-/**
- * MCP Server Integration Tests.
- *
- * Business Context: MCP protocol compliance is essential for AI agent integration.
- *
- * Decision Rationale: Test tool handler and error cases that can be validated
- * without requiring a connected transport.
- */
-describe("MCP Server Integration", () => {
-  let server: ReturnType<typeof createServer>;
-
-  beforeEach(() => {
-    server = createServer();
-  });
-
-  it("server can be created without errors", () => {
-    expect(server).toBeDefined();
-    expect(typeof server.connect).toBe("function");
-    expect(typeof server.close).toBe("function");
-  });
-
-  it("rejects unknown tool name", async () => {
-    const request = {
-      method: "tools/call" as const,
-      params: {
-        name: "unknownTool",
-        arguments: {}
-      }
-    };
-
-    await expect(server.request(request, CallToolRequestSchema)).rejects.toThrow();
-  });
-
-  it("handles valid decision analysis request", () => {
-    const validInput = {
-      decisionStatement: "Should we implement the new feature?",
-      options: [
-        { id: "opt1", name: "Implement Feature", description: "Add the new feature" },
-        { id: "opt2", name: "Delay Feature", description: "Wait for more data" }
-      ],
-      stakeholders: ["Product Manager"],
-      constraints: [],
-      timeHorizon: "6 months",
-      riskTolerance: "risk-neutral",
-      decisionId: "feature-001",
-      analysisType: "multi-criteria",
-      stage: "options",
-      iteration: 1,
-      nextStageNeeded: false
-    };
-
-    const result = validateDecisionAnalysisDataForTest(validInput);
-    expect(result.decisionStatement).toBe("Should we implement the new feature?");
-    expect(result.options).toHaveLength(2);
-  });
-});
-
-/**
- * Edge Cases and Performance Tests.
- *
- * Business Context: Enterprise applications must handle edge cases gracefully
- * and perform well under load.
- *
- * Decision Rationale: Test boundary conditions and performance to ensure
- * production reliability.
- */
-describe("Edge Cases and Performance", () => {
-  it("handles large number of options", () => {
-    const manyOptions = Array.from({ length: 50 }, (_, i) => ({
-      id: `opt${i + 1}`,
-      name: `Option ${i + 1}`,
-      description: `Description for option ${i + 1}`
-    }));
-
-    const input = {
-      decisionStatement: "Decision with many options",
-      options: manyOptions,
-      stakeholders: ["Test"],
-      constraints: [],
-      timeHorizon: "1 year",
-      riskTolerance: "risk-neutral",
-      decisionId: "many-options-test",
-      analysisType: "multi-criteria",
-      stage: "options",
-      iteration: 1,
-      nextStageNeeded: false
-    };
-
-    const result = validateDecisionAnalysisDataForTest(input);
-    expect(result.options).toHaveLength(50);
-  });
-
-  it("handles complex nested structures", () => {
-    const complexInput = {
-      decisionStatement: "Complex decision with nested evaluations",
-      options: [{ id: "opt1", name: "Option 1", description: "Complex option" }],
-      criteria: [
-        {
-          id: "criterion1",
-          name: "Complex Criterion",
-          description: "A criterion with complex evaluation",
-          weight: 1.0,
-          evaluationMethod: "quantitative"
-        }
-      ],
-      criteriaEvaluations: [
-        {
-          criterionId: "criterion1",
-          optionId: "opt1",
-          score: 0.85,
-          justification: "Complex justification with detailed reasoning and multiple factors to consider"
-        }
-      ],
-      possibleOutcomes: [
-        {
-          id: "outcome1",
-          description: "Complex outcome with detailed description",
-          probability: 0.75,
-          optionId: "opt1",
-          value: 500000,
-          confidenceInEstimate: 0.8
-        }
-      ],
-      informationGaps: [
-        {
-          description: "Complex information gap requiring extensive research",
-          impact: 0.9,
-          researchMethod: "Comprehensive market analysis with stakeholder interviews"
-        }
-      ],
-      stakeholders: ["Complex Stakeholder 1", "Complex Stakeholder 2"],
-      constraints: ["Complex constraint 1", "Complex constraint 2"],
-      timeHorizon: "3 years",
-      riskTolerance: "risk-seeking",
-      decisionId: "complex-test",
-      analysisType: "multi-criteria",
-      stage: "analysis",
-      iteration: 3,
-      nextStageNeeded: true,
-      suggestedNextStage: "recommendation"
-    };
-
-    const result = validateDecisionAnalysisDataForTest(complexInput);
-    expect(result.criteriaEvaluations[0].justification).toContain("Complex justification");
-    expect(result.possibleOutcomes[0].description).toContain("Complex outcome");
-    expect(result.informationGaps[0].researchMethod).toContain("Comprehensive market analysis");
-  });
-
-  it("handles zero iteration", () => {
-    const input = {
-      decisionStatement: "Initial decision",
-      options: [{ id: "opt1", name: "Option 1", description: "Test option" }],
-      stakeholders: ["Test"],
-      constraints: [],
-      timeHorizon: "1 year",
-      riskTolerance: "risk-neutral",
-      decisionId: "zero-iteration-test",
-      analysisType: "multi-criteria",
-      stage: "problem-definition",
-      iteration: 0,
-      nextStageNeeded: true
-    };
-
-    const result = validateDecisionAnalysisDataForTest(input);
-    expect(result.iteration).toBe(0);
-  });
-
-  it("handles empty arrays", () => {
-    const input = {
-      decisionStatement: "Decision with empty arrays",
-      options: [{ id: "opt1", name: "Option 1", description: "Test option" }],
-      criteria: [],
-      criteriaEvaluations: [],
-      possibleOutcomes: [],
-      informationGaps: [],
-      stakeholders: ["Test"],
-      constraints: [],
-      timeHorizon: "1 year",
-      riskTolerance: "risk-neutral",
-      decisionId: "empty-arrays-test",
-      analysisType: "multi-criteria",
-      stage: "options",
-      iteration: 1,
-      nextStageNeeded: false
-    };
-
-    const result = validateDecisionAnalysisDataForTest(input);
-    expect(result.criteria || []).toHaveLength(0);
-    expect(result.criteriaEvaluations || []).toHaveLength(0);
-    expect(result.possibleOutcomes || []).toHaveLength(0);
-    expect(result.informationGaps || []).toHaveLength(0);
+    expect(result.multiCriteriaScores).toBeDefined();
+    expect(result.multiCriteriaScores?.opt1).toBeCloseTo(0.8);
   });
 });
