+++ src/decision-framework/tsconfig.build.json
@@ -16,5 +16,5 @@
   },
 
   "include": ["./src/**/*.ts"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "src/tests"]
 }
